deps/v8/test/webkit/fast/js/Promise-resolve-with-itself-expected.txt

Test whether Promise will be rejected if it is resolved with itself.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rejected
PASS result is "TypeError: Chaining cycle detected for promise #<Promise>"
PASS successfullyParsed is true

TEST COMPLETE
